android/client/app/src/main/java/id/ac/ui/cs/mobileprogramming/ilham_darmawan/konselink_client/model/room/UserDao.kt
package id.ac.ui.cs.mobileprogramming.ilham_darmawan.konselink_client.model.room

import androidx.lifecycle.LiveData
import androidx.room.Dao
import androidx.room.Delete
import androidx.room.Insert
import androidx.room.Query

@Dao
interface UserDao {
    @Query("SELECT * FROM users")
    fun getAllUsers(): List<User>

    @Query("SELECT * FROM users ORDER BY id DESC LIMIT 1")
    fun getLatestUser(): LiveData<User>

    @Query("SELECT * FROM users where academicId = :academicId")
    fun getUserById(academicId: String): List<User>

    @Insert
    fun insert(user: User)

    @Delete
    fun delete(user: User)

    @Query("DELETE FROM users")
    fun deleteAll()
}